requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.12.5
pydantic-settings==2.11.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
authlib==1.3.0
httpx==0.26.0
openai==1.12.0
cryptography==42.0.2
stripe==8.3.0
